PSR-787 Add error message for files that without .csv extension, make uploading files generic and configure spinners
